requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
mcp>=0.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
requests>=2.31.0
